Move desrialize to utils; Remove converter from old research
